+++ app/components/edit/edit.tsx
@@ -1,0 +1,28 @@
+"use client";
+
+
+import EditPageComponent from "@/app/components/edit/edit-page";
+import { EventModalProps } from "@/app/lib/types";
+
+
+// No need for suspense, EditPageComponent already handles loading and edge cases
+export default function EditPage({ event, onClose } : EventModalProps) {
+
+    return (
+        <main className="min-h-screen w-screen bg-gradient-to-b from-[#083157]  to-[#064580]">
+            <EditPageComponent event={event} onClose={onClose} />
+        </main>
+    )
+
+}
+
+
+// function LoadingScreen() {
+// 	return (
+// 		<main className="min-h-screen w-screen bg-gradient-to-b from-[#083157]  to-[#064580]">
+// 			<div className="flex flex-col items-center justify-center">
+// 				<h1 className="text-3xl">Loading...</h1>
+// 			</div>
+// 		</main>
+// 	)
+// }