+++ app/events/edit/page.tsx
@@ -1,27 +1,28 @@
-"use client";
+// "use client";
 
-import { Suspense } from "react";
-import EditPageComponent from "@/app/components/edit/edit-page";
 
+// import EditPageComponent from "@/app/components/edit/edit-page";
+// import { EventModalProps } from "@/app/lib/types";
+
+
+// // No need for suspense, EditPageComponent already handles loading and edge cases
 export default function EditPage() {
 
-	return (
-		<main className="min-h-screen w-screen bg-gradient-to-b from-[#083157]  to-[#064580]">
-				<Suspense fallback={<LoadingScreen />} >
-					<EditPageComponent />
-				</Suspense>
-			</main>
+	return (<></>
+// 		<main className="min-h-screen w-screen bg-gradient-to-b from-[#083157]  to-[#064580]">
+// 			<EditPageComponent event={event} onClose={onClose} />
+// 		</main>
 	)
 
-}
+// }
 
 
-function LoadingScreen() {
-	return (
-		<main className="min-h-screen w-screen bg-gradient-to-b from-[#083157]  to-[#064580]">
-			<div className="flex flex-col items-center justify-center">
-				<h1 className="text-3xl">Loading...</h1>
-			</div>
-		</main>
-	)
+// function LoadingScreen() {
+// 	return (
+// 		<main className="min-h-screen w-screen bg-gradient-to-b from-[#083157]  to-[#064580]">
+// 			<div className="flex flex-col items-center justify-center">
+// 				<h1 className="text-3xl">Loading...</h1>
+// 			</div>
+// 		</main>
+// 	)
 }